+++ src/cobib/man/__init__.py
@@ -2,3 +2,5 @@
 
 .. include:: cobib.1.html_fragment
 """
+
+from .manual import manual as manual