+++ src/cobib/ui/components/__init__.py
@@ -14,6 +14,7 @@ from .list_view import ListView as ListView
 from .log_screen import LogScreen as LogScreen
 from .logging_handler import LoggingHandler as LoggingHandler
 from .main_content import MainContent as MainContent
+from .manual_screen import ManualScreen as ManualScreen
 from .motion_key import MotionKey as MotionKey
 from .note_view import NoteView as NoteView
 from .preset_filter_screen import PresetFilterScreen as PresetFilterScreen